GridCreatingConditionCompoundChannel::SwitchStatusCommand separated
