+++ libs/gridcreatingcondition/compoundchannel/gridcreatingconditioncompoundchannel.cpp
@@ -10,6 +10,7 @@
 
 #include "private/gridcreatingconditioncompoundchannel_polygoncoordinateseditor.h"
 #include "private/gridcreatingconditioncompoundchannel_polylinecoordinateseditor.h"
+#include "private/gridcreatingconditioncompoundchannel_switchstatuscommand.h"
 
 #include <guibase/widget/waitdialog.h>
 #include <guicore/base/iricmainwindowinterface.h>
@@ -179,79 +180,6 @@ void GridCreatingConditionCompoundChannel::viewOperationEnded(PreProcessorGraphi
 {
 	updateMouseCursor(v);
 }
-
-class GridCreatingConditionCompoundChannel::SwitchStatusCommand : public QUndoCommand
-{
-public:
-	SwitchStatusCommand(GridCreatingConditionCompoundChannel::Status newStatus, GridCreatingConditionCompoundChannel* polygon) :
-		QUndoCommand {GridCreatingConditionCompoundChannel::tr("Finish Defining Polygon or Polygonal line")}
-	{
-		m_newStatus = newStatus;
-		m_condition = polygon;
-	}
-	void undo() {
-		switch (m_newStatus) {
-		case GridCreatingConditionCompoundChannel::stDefiningLowWaterRegion:
-			m_condition->m_status = GridCreatingConditionCompoundChannel::stDefiningRegion;
-			m_condition->m_selectedPolygon = m_condition->m_gridRegionPolygon;
-			m_condition->m_gridRegionPolygon->setActive(true);
-			m_condition->m_lowWaterChannelPolygon->setActive(false);
-			break;
-		case GridCreatingConditionCompoundChannel::stDefiningCenterLine:
-			m_condition->m_status = GridCreatingConditionCompoundChannel::stDefiningLowWaterRegion;
-			m_condition->m_selectedPolygon = m_condition->m_lowWaterChannelPolygon;
-			m_condition->m_lowWaterChannelPolygon->setActive(true);
-			m_condition->m_centerLine->setActive(false);
-			break;
-		case GridCreatingConditionCompoundChannel::stNormal:
-			m_condition->m_status = GridCreatingConditionCompoundChannel::stDefiningCenterLine;
-			m_condition->m_selectedLine = m_condition->m_centerLine;
-			m_condition->m_centerLine->setActive(true);
-			break;
-		default:
-			break;
-		}
-		m_condition->m_mouseEventMode = GridCreatingConditionCompoundChannel::meDefining;
-
-		m_condition->updateMouseCursor(m_condition->graphicsView());
-		m_condition->updateActionStatus();
-		m_condition->renderGraphicsView();
-	}
-	void redo() {
-		m_condition->m_status = m_newStatus;
-		switch (m_newStatus) {
-		case GridCreatingConditionCompoundChannel::stDefiningLowWaterRegion:
-			m_condition->m_mouseEventMode = GridCreatingConditionCompoundChannel::meBeforeDefining;
-			m_condition->m_selectedPolygon = m_condition->m_lowWaterChannelPolygon;
-			m_condition->m_gridRegionPolygon->setActive(false);
-			m_condition->m_lowWaterChannelPolygon->setActive(true);
-			InformationDialog::information(m_condition->preProcessorWindow(), GridCreatingConditionCompoundChannel::tr("Information"), GridCreatingConditionCompoundChannel::tr("Next, please define low water channel region. Water channel can be defined as polygon by mouse-clicking. Finish definining by double clicking, or pressing return key."), "gccompoundchannel_lowwater");
-			break;
-		case GridCreatingConditionCompoundChannel::stDefiningCenterLine:
-			m_condition->m_mouseEventMode = GridCreatingConditionCompoundChannel::meBeforeDefining;
-			m_condition->m_selectedLine = m_condition->m_centerLine;
-			m_condition->m_lowWaterChannelPolygon->setActive(false);
-			m_condition->m_centerLine->setActive(true);
-			InformationDialog::information(m_condition->preProcessorWindow(), GridCreatingConditionCompoundChannel::tr("Information"), GridCreatingConditionCompoundChannel::tr("Next, please define grid center line. Grid center line can be defined as polygonal line by mouse-clicking. Finish definining by double clicking, or pressing return key."), "gccompoundchannel_centerline");
-			break;
-		case GridCreatingConditionCompoundChannel::stNormal:
-			m_condition->m_mouseEventMode = GridCreatingConditionCompoundChannel::meNormal;
-			m_condition->m_selectMode = GridCreatingConditionCompoundChannel::smLine;
-			m_condition->m_selectedLine = m_condition->m_centerLine;
-			m_condition->m_centerLine->setActive(true);
-			m_condition->create(m_condition->preProcessorWindow());
-			break;
-		default:
-			break;
-		}
-		m_condition->updateMouseCursor(m_condition->graphicsView());
-		m_condition->updateActionStatus();
-		m_condition->renderGraphicsView();
-	}
-private:
-	GridCreatingConditionCompoundChannel::Status m_newStatus;
-	GridCreatingConditionCompoundChannel* m_condition;
-};
 
 void GridCreatingConditionCompoundChannel::keyPressEvent(QKeyEvent* event, PreProcessorGraphicsViewInterface* /*v*/)
 {
@@ -1408,9 +1336,9 @@ void GridCreatingConditionCompoundChannel::definePolygon(bool doubleClick)
 	stack.beginMacro(tr("Finish Defining Polygon"));
 	// finish defining the polygon.
 	if (m_status == stDefiningRegion) {
-		stack.push(new SwitchStatusCommand(stDefiningLowWaterRegion, this));
+		pushRenderCommand(new SwitchStatusCommand(stDefiningLowWaterRegion, this));
 	} else if (m_status == stDefiningLowWaterRegion) {
-		stack.push(new SwitchStatusCommand(stDefiningCenterLine, this));
+		pushRenderCommand(new SwitchStatusCommand(stDefiningCenterLine, this));
 	}
 	stack.endMacro();
 }
@@ -1430,7 +1358,7 @@ void GridCreatingConditionCompoundChannel::defineLine(bool doubleClick)
 	stack.undo();
 	stack.beginMacro(tr("Finish Defining Polygonal line"));
 	// finish defining the line.
-	stack.push(new SwitchStatusCommand(stNormal, this));
+	pushRenderCommand(new SwitchStatusCommand(stNormal, this));
 	stack.endMacro();
 }
